+++ lib/plugins/sfDoctrinePlugin/config/sfDoctrinePluginConfiguration.class.php
@@ -37,11 +37,18 @@ class sfDoctrinePluginConfiguration extends sfPluginConfiguration
       $this->dispatcher->connect('debug.web.load_panels', array('sfWebDebugPanelDoctrine', 'listenToAddPanelEvent'));
     }
 
-    if (!class_exists('Doctrine_Core', false))
+    if (file_exists($compiledFile = sfConfig::get('sf_cache_dir') . '/doctrine.compiled.php'))
     {
-      require_once sfConfig::get('sf_doctrine_dir', realpath(dirname(__FILE__).'/../lib/vendor/doctrine/lib')).'/Doctrine/Core.php';
+      require_once $compiledFile;
     }
-    spl_autoload_register(array('Doctrine_Core', 'autoload'));
+    else
+    {
+      if (!class_exists('Doctrine_Core', false))
+      {
+        require_once sfConfig::get('sf_doctrine_dir', realpath(dirname(__FILE__).'/../lib/vendor/doctrine/lib')).'/Doctrine/Core.php';
+      }
+      spl_autoload_register(array('Doctrine_Core', 'autoload'));
+    }
 
     $manager = Doctrine_Manager::getInstance();
     $manager->setAttribute(Doctrine_Core::ATTR_EXPORT, Doctrine_Core::EXPORT_ALL);